[Feat] #2 PlaceViewBeforeCheckInUI
