+++ BNomad/View/PlaceViewBeforeCheckIn/CustomCollectionViewCell.swift
@@ -1,9 +1,8 @@
 //
-//  CustomCollectionViewCell.swift
-//  BottomSheet
+//  CollectionViewCell.swift
+//  BNomad
 //
 //  Created by 박진웅 on 2022/10/18.
-//  Copyright © 2022 Zafar. All rights reserved.
 //
 
 import UIKit
@@ -12,27 +11,7 @@ class CollectionViewCell: UICollectionViewCell {
 
     static let identifier = "CustomCollectionViewCell"
 
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-//    required init?(coder: NSCoder) {
-//        super.init(coder: coder)
-//        setUpCell()
-//    }
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setUpCell()
-    }
-
-    func setUpCell() {
-        contentView.addSubview(cell)
-        cell.centerXAnchor.constraint(equalTo: contentView.centerXAnchor).isActive = true
-        cell.widthAnchor.constraint(equalToConstant: 356).isActive = true
-        cell.heightAnchor.constraint(equalToConstant: 86).isActive = true
-    }
-    
+    // MARK: - Properties
     lazy var cell: UIButton = {
         let rectangle = UIButton()
         rectangle.frame = CGRect(x: 0, y: 0, width: 356, height: 86)
@@ -40,18 +19,14 @@ class CollectionViewCell: UICollectionViewCell {
         rectangle.translatesAutoresizingMaskIntoConstraints = false
         rectangle.backgroundColor = .white
         rectangle.addSubview(name)
-        name.topAnchor.constraint(equalTo: rectangle.topAnchor, constant: 23).isActive = true
-        name.leftAnchor.constraint(equalTo: rectangle.leftAnchor, constant: 9).isActive = true
         rectangle.addSubview(numberOfCheckIn)
-        numberOfCheckIn.topAnchor.constraint(equalTo: rectangle.topAnchor, constant: 26).isActive = true
-        numberOfCheckIn.rightAnchor.constraint(equalTo: rectangle.rightAnchor, constant: -50).isActive = true
         rectangle.addSubview(distance)
-        distance.bottomAnchor.constraint(equalTo: rectangle.bottomAnchor, constant: -8).isActive = true
-        distance.leftAnchor.constraint(equalTo: rectangle.leftAnchor, constant: 11).isActive = true
         rectangle.addSubview(averageTime)
-        averageTime.bottomAnchor.constraint(equalTo: rectangle.bottomAnchor, constant: -8).isActive = true
-        averageTime.rightAnchor.constraint(equalTo: rectangle.rightAnchor, constant: -48).isActive = true
         rectangle.addSubview(arrow)
+        name.anchor(top: rectangle.topAnchor, left: rectangle.leftAnchor, paddingTop: 23, paddingLeft: 9)
+        numberOfCheckIn.anchor(top: rectangle.topAnchor, right: rectangle.rightAnchor, paddingTop: 26, paddingRight: 50)
+        distance.anchor(left: rectangle.leftAnchor, bottom: rectangle.bottomAnchor, paddingLeft: 11, paddingBottom: 8)
+        averageTime.anchor(bottom: rectangle.bottomAnchor, right: rectangle.rightAnchor, paddingBottom: 8, paddingRight: 48)
         arrow.centerYAnchor.constraint(equalTo: rectangle.centerYAnchor).isActive = true
         arrow.rightAnchor.constraint(equalTo: rectangle.rightAnchor, constant: -12.45).isActive = true
         return rectangle
@@ -61,9 +36,8 @@ class CollectionViewCell: UICollectionViewCell {
         let title = UILabel()
         title.backgroundColor = .clear
         title.textColor = .black
-//        title.font = UIFont.preferredFont(forTextStyle: UIFont.TextStyle.subheadline)
+        title.font = .preferredFont(forTextStyle: .title2, weight: .bold)
         title.text = "투썸 플레이스"
-        title.font = UIFont(name: "SFProDisplay-Bold", size: 22)
         title.textAlignment = .center
         title.translatesAutoresizingMaskIntoConstraints = false
         return title
@@ -73,7 +47,7 @@ class CollectionViewCell: UICollectionViewCell {
         let title = UILabel()
         title.backgroundColor = .clear
         title.textColor = .black
-//        title.font = UIFont.preferredFont(forTextStyle: UIFont.TextStyle.subheadline)
+        title.font = .preferredFont(forTextStyle: .body, weight: .regular)
         title.text = "3명 체크인"
         title.font = UIFont(name: "SFProText-Regular", size: 17)
         title.textAlignment = .center
@@ -84,8 +58,8 @@ class CollectionViewCell: UICollectionViewCell {
     var distance: UILabel = {
         let title = UILabel()
         title.backgroundColor = .clear
-        title.textColor = UIColor(red: 0.557, green: 0.557, blue: 0.576, alpha: 1)
-//        title.font = UIFont.preferredFont(forTextStyle: UIFont.TextStyle.subheadline)
+        title.textColor = CustomColor.nomadGray2
+        title.font = .preferredFont(forTextStyle: .footnote, weight: .regular)
         title.text = "1.5km"
         title.font = UIFont(name: "SFProText-Regular", size: 15)
         title.textAlignment = .center
@@ -96,8 +70,8 @@ class CollectionViewCell: UICollectionViewCell {
     var averageTime: UILabel = {
         let title = UILabel()
         title.backgroundColor = .clear
-        title.textColor = UIColor(red: 0.557, green: 0.557, blue: 0.576, alpha: 1)
-//        title.font = UIFont.preferredFont(forTextStyle: UIFont.TextStyle.subheadline)
+        title.textColor = CustomColor.nomadGray2
+        title.font = .preferredFont(forTextStyle: .footnote, weight: .regular)
         title.text = "평균 5시간 근무"
         title.font = UIFont(name: "SFProText-Regular", size: 13)
         title.textAlignment = .center
@@ -107,9 +81,25 @@ class CollectionViewCell: UICollectionViewCell {
     
     let arrow: UIImageView = {
 //        let config = UIImage.SymbolConfiguration(pointSize: 16.95)
-        let image = UIImage(systemName: "chevron.right")?.withTintColor(.blue, renderingMode: .alwaysOriginal)
+        let image = UIImage(systemName: "chevron.right")?.withTintColor(CustomColor.nomadBlue ?? .blue, renderingMode: .alwaysOriginal)
         let imageView = UIImageView(image: image)
         imageView.translatesAutoresizingMaskIntoConstraints = false
         return imageView
     }()
+    
+    // MARK: - LifeCycle
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        setUpCell()
+    }
+
+    func setUpCell() {
+        contentView.addSubview(cell)
+        cell.anchor(width: 356, height: 86)
+        cell.centerXAnchor.constraint(equalTo: contentView.centerXAnchor).isActive = true
+    }
 }
