+++ BNomad/View/PlaceViewBeforeCheckIn/FinalViewController.swift
@@ -1,0 +1,21 @@
+//
+//  FinalViewController.swift
+//  BNomad
+//
+//  Created by 박진웅 on 2022/10/19.
+//
+//  Copyright © 2020 Zafar. All rights reserved.
+//
+
+import UIKit
+
+final class FinalViewController: ModalViewController
+<MapViewController, CustomModalViewController> {
+    
+    override func viewDidLoad() {
+        
+        super.viewDidLoad()
+        
+    }
+    
+}
